Use the cake types that are defined in the cake qdisc
While json sends everything around as a u64, many of the cake
variables internally and in the protocol, are 16-32 bit. It
seemed a major memory savings to have the types more closely
match the underlying reality, and one small step towards
actually getting these values from the kernel more directly.

In terms of surprise several of those can actually roll over,
which is something we need to look out for in the future.

+++ src/rust/lqosd/src/queue_tracker/queue_reader/queue_diff/mod.rs
@@ -25,18 +25,18 @@ pub(crate) fn make_queue_diff(previous: &QueueType, current: &QueueType) -> Resu
 #[derive(Serialize, Clone, Debug)]
 pub struct CakeDiff {
     pub bytes: u64,
-    pub packets: u64,
-    pub qlen: u64,
+    pub packets: u32,
+    pub qlen: u32,
     pub tins: Vec<CakeDiffTin>,
 }
 
 #[derive(Serialize, Clone, Debug)]
 pub struct CakeDiffTin {
     pub sent_bytes: u64,
-    pub backlog_bytes: u64,
-    pub drops: u64,
-    pub marks: u64,
-    pub avg_delay_us: u64,
+    pub backlog_bytes: u32,
+    pub drops: u32,
+    pub marks: u32,
+    pub avg_delay_us: u32,
 }
 
 fn cake_diff(previous: &QueueType, current: &QueueType) -> Result<QueueDiff> {
